fix(doc-gate): 排除名单统一至 _doc_gate_common + .claude/logs 降噪 (#23)
抽取共享模块 _doc_gate_common.py 作为排除名单与链接图谱原语的单一来源，
recall-gate.py 与 docs-graph.py 改为 import，根除三处硬编码漂移。

- EXCLUDED_DIRS 新增 .claude、logs；recall-gate.sh Phase 6 补 */logs/*
  （sh 写时门禁与 Python 语料库扫描是两套独立机制，须同时覆盖）
- 删除 archive 豁免死规则 ORPHAN_WHITELIST_DIRS（v1.1.0 从 Infra 迁入、本项目零覆盖）
- docs-graph extract_links 降为薄包装，errors='replace' 保留
- 顺删抽取后 unused import（recall: urllib.parse；docs-graph: re/os/urllib.parse）
- 版本 1.2.2 → 1.3.0（plugin.json + marketplace.json 双写）

净 -172 行，去重收敛。

测试（私有套件）全绿：test_doc_gate_common(17) + test_recall_gate(67)
+ recall-gate.bats(40) + test-docs-graph(28)。
顺修 #22 遗留的 2 个 detect_root 测试债（语义未随 innermost→outermost 同步）。
