+++ .claude-plugin/marketplace.json
@@ -30,7 +30,7 @@
     {
       "name": "doc-gate",
       "description": "Document editing governance — doc-maintenance skill + hard gate + recall advisory",
-      "version": "1.2.2",
+      "version": "1.3.0",
       "author": {
         "name": "WooDragon"
       },
